# Use local variables for stack on JIT-ed method